+++ helper/uuids.js
@@ -1,0 +1,4 @@
+module.exports = () =>
+  Math.floor((1 + Math.random()) * 0x10000)
+    .toString(16)
+    .substring(1);
